feat: add URL security protections and Neon environment configuration
- add safe URL validations to the shortening flow
- prepare the foundation for suspicious URL blocking/control
- update environment variables for Neon usage (dev/prod)
- improve compose configuration to switch env files

+++ url-shortering/src/main/java/br/com/systemdesign/urlshortening/dto/CreateShortUrlRequest.java
@@ -9,5 +9,5 @@ public record CreateShortUrlRequest(
     @URL(message = "{validation.url.original.invalid}")
     @SafeUrl
     String originalUrl,
-    Integer expirationSeconds
+    Long expirationSeconds
 ) {}
